+++ packages/facets/tsconfig-json/src/features/detect.ts
@@ -1,6 +1,7 @@
 import path from 'node:path'
 
 import { FacetDetectionResult, DetectFunction, WhimbrelContext } from '@whimbrel/core-api'
+import { TsConfigJSON } from '@src/adapters'
 
 export const detect: DetectFunction = async (
   ctx: WhimbrelContext,
@@ -9,6 +10,8 @@ export const detect: DetectFunction = async (
   const tsConfigJsonPath = path.join(dir, 'tsconfig.json')
 
   if (await ctx.disk.exists(tsConfigJsonPath)) {
+    const tree = await TsConfigJSON.readReferenceTree(ctx.disk, tsConfigJsonPath)
+
     return {
       detected: true,
       facet: {
@@ -16,6 +19,7 @@ export const detect: DetectFunction = async (
           roles: ['build-config'],
           config: {
             path: tsConfigJsonPath,
+            configTree: tree,
           },
         },
       },