+++ packages/facets/tsconfig-json/src/query/version-control.ignore-files.ts
@@ -1,6 +1,61 @@
 import path from 'node:path'
 import { FacetQueryFunction, WhimbrelContext } from '@whimbrel/core-api'
-import { readPath } from '@whimbrel/walk'
+import { TsConfigJSON } from '@src/adapters'
+import { ReferenceTreeNode } from '@src/adapters/tsconfig.json-adapter'
+import { pushUnique } from '@whimbrel/array'
+
+/**
+ * Extract ignore file patterns from a tsconfig.json file.
+ *
+ * @param file - The TsConfigJSON file to extract ignore patterns from.
+ * @param root - The root directory of the actor.
+ *
+ * @return An array of ignore file patterns.
+ */
+const extractIgnoreFiles = (file: TsConfigJSON, root: string) => {
+  const files = []
+
+  const outDir = file.get('compilerOptions.outDir')
+
+  if (outDir) {
+    files.push({ pattern: `${outDir}/`, groups: ['build'], source: 'tsconfig.json' })
+  }
+
+  const tsBuildInfoPath = file.get<string>('compilerOptions.tsBuildInfoFile')
+  if (tsBuildInfoPath) {
+    files.push({ pattern: tsBuildInfoPath, groups: ['build'], source: 'tsconfig.json' })
+  } else if (!file.get('references')) {
+    files.push({
+      pattern: `${path.parse(path.basename(file.getPath())).name}.tsbuildinfo`,
+      groups: ['build'],
+      source: 'tsconfig.json',
+    })
+  }
+
+  return files
+}
+
+/**
+ * Recursively walk the tsconfig.json reference tree to extract ignore files.
+ *
+ * @param refTree - The reference tree node to walk.
+ * @param root - The root directory of the actor.
+ */
+const walkReferenceTree = (refTree: ReferenceTreeNode, root: string) => {
+  if ('recursive' in refTree) return []
+
+  const files = extractIgnoreFiles(refTree.file!, root)
+
+  for (const ref of refTree.references) {
+    pushUnique(files, ...walkReferenceTree(ref, root))
+  }
+
+  if (refTree.extends) {
+    pushUnique(files, ...walkReferenceTree(refTree.extends, root))
+  }
+
+  return files
+}
 
 /**
  * Query implementation of `version-control:ignore-files`.
@@ -22,18 +77,10 @@ export const queryVersionControlIgnoreFiles: FacetQueryFunction<
     const tsConfigPath =
       tsConfigScope?.config?.path ?? path.join(actor.root, 'tsconfig.json')
 
-    if (await ctx.disk.exists(tsConfigPath)) {
-      const tsConfigJson = await ctx.disk.readJson(tsConfigPath)
-      const outDir = readPath(tsConfigJson, 'compilerOptions.outDir')
-      if (outDir) {
-        return [
-          {
-            pattern: `${path.basename(outDir)}/`,
-            groups: ['build'],
-            source: 'tsconfig.json',
-          },
-        ]
-      }
-    }
+    const refTree = await TsConfigJSON.readReferenceTree(ctx.disk, tsConfigPath, true)
+
+    const ignoreFiles = walkReferenceTree(refTree, actor.root)
+
+    return ignoreFiles
   }
 }